Trainer: Refactoring the clava_synthesizing function (#78)
Breaking down the clava_synthesizing function of the models/clavaddpm/synthesizer.py module into 3 functions: one for the single table synthesizing, one for the multi table synthesizing and one for the post processing of the data.

Also did a little refactor in the types to include the Relation and GroupLengthProbDict in order to simplify some types we have in our code. I have added a note on the ticket to later change in other places where those types are present.